# arch-studio-source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Architecture Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, category, description, imageUrl, featured, year]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [residential, commercial, interior, landscape, renovation]
        description:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        year:
          type: integer
        location:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
...